Remove extra fragments adn resources associated with main graph
