+++ app/src/main/java/com/jp_funda/todomind/view/mind_map_create/options_sheet/MindMapOptionsSheet.kt
@@ -1,9 +1,5 @@
 package com.jp_funda.todomind.view.mind_map_create.options_sheet
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.ExitTransition
 import androidx.compose.animation.ExperimentalAnimationApi
@@ -13,156 +9,178 @@ import androidx.compose.animation.slideInHorizontally
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
-import androidx.fragment.app.activityViewModels
-import androidx.fragment.app.viewModels
+import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.pager.ExperimentalPagerApi
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.jp_funda.todomind.R
+import com.jp_funda.todomind.data.repositories.task.entity.Task
 import com.jp_funda.todomind.data.repositories.task.entity.TaskStatus
 import com.jp_funda.todomind.view.MainViewModel
 import com.jp_funda.todomind.view.components.MindMapOptionsTabRow
 import com.jp_funda.todomind.view.components.TaskEditContent
 import com.jp_funda.todomind.view.mind_map_create.MindMapCreateViewModel
-import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.launch
 
-@ExperimentalAnimationApi
 @ExperimentalMaterialApi
 @ExperimentalPagerApi
-@AndroidEntryPoint
-class MindMapOptionsSheet : BottomSheetDialogFragment() {
+@ExperimentalAnimationApi
+@Composable
+fun MindMapOptionsSheet(bottomSheetState: BottomSheetState, mainViewModel: MainViewModel) {
+    val addChildViewModel = hiltViewModel<AddChildViewModel>()
+    val sheetViewModel = hiltViewModel<MindMapOptionsSheetViewModel>()
 
-    // ViewModels
-    private val sheetViewModel by viewModels<MindMapOptionsSheetViewModel>()
-    private val addChildViewModel by viewModels<AddChildViewModel>()
-    private val editTaskViewModel by viewModels<EditTaskViewModel>()
-    private val mindMapCreateViewModel by activityViewModels<MindMapCreateViewModel>()
-    private val mainViewModel by activityViewModels<MainViewModel>()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        setUpAddingChildNode()
-
+    LaunchedEffect(bottomSheetState.isCollapsed) {
+        // Set Sheet Mode as Edit
+        sheetViewModel.setMode(MindMapOptionsMode.EDIT_TASK)
         // Set mind map to addChildViewModel
         mainViewModel.editingMindMap?.let { addChildViewModel.setMindMap(it) }
         // Set parentTask to addChildViewModel
-        mainViewModel.selectedNode?.let { addChildViewModel.initializeParentTask(it) }
+        sheetViewModel.selectedNode.value?.let { addChildViewModel.initializeParentTask(it) }
+        // Set up node at addChildViewModel
+        setUpAddingChildNode(sheetViewModel.selectedNode.value, mainViewModel, addChildViewModel)
         // Set addChildViewModel's new task status as open
         addChildViewModel.setStatus(TaskStatus.Open)
+    }
 
-        return ComposeView(requireContext()).apply {
-            setContent {
-                val observedMode = sheetViewModel.selectedMode.observeAsState()
-                observedMode.value?.let { selectedMode ->
-                    Column(
-                        modifier = Modifier
-                            .padding(bottom = 20.dp)
-                            .background(colorResource(id = R.color.deep_purple)),
+    val observedNode = sheetViewModel.selectedNode.observeAsState()
+    observedNode.value.run {
+        MindMapOptionsSheetContent(
+            bottomSheetState = bottomSheetState,
+            mainViewModel = mainViewModel,
+        )
+    }
+}
+
+@ExperimentalMaterialApi
+@ExperimentalPagerApi
+@ExperimentalAnimationApi
+@Composable
+fun MindMapOptionsSheetContent(bottomSheetState: BottomSheetState, mainViewModel: MainViewModel) {
+    val sheetViewModel = hiltViewModel<MindMapOptionsSheetViewModel>()
+    val addChildViewModel = hiltViewModel<AddChildViewModel>()
+    val editTaskViewModel = hiltViewModel<EditTaskViewModel>()
+    val mindMapCreteViewModel = hiltViewModel<MindMapCreateViewModel>()
+
+    val coroutineScope = rememberCoroutineScope()
+    val sheetHeight = LocalConfiguration.current.screenHeightDp * 0.7
+
+    val observedMode = sheetViewModel.selectedMode.observeAsState()
+
+    observedMode.value?.let { selectedMode ->
+        Column(
+            modifier = Modifier
+                .padding(bottom = 20.dp)
+                .background(colorResource(id = R.color.deep_purple))
+                .height(sheetHeight.dp)
+        ) {
+            // When taskNode is Selected
+            if (mainViewModel.selectedNode != null) {
+                MindMapOptionsTabRow(selectedMode = selectedMode) {
+                    sheetViewModel.setMode(it)
+                }
+
+                // Edit Task Option
+                // set Editing Task
+                mainViewModel.selectedNode?.let { editTaskViewModel.setEditingTask(it) }
+                AnimatedVisibility(
+                    visible = selectedMode == MindMapOptionsMode.EDIT_TASK,
+                    enter = slideInHorizontally(
+                        initialOffsetX = { -300 }, // small slide 300px
+                        animationSpec = tween(
+                            durationMillis = 200,
+                            easing = LinearEasing // interpolator
+                        )
+                    ),
+                    exit = ExitTransition.None
+                ) {
+                    TaskEditContent(
+                        taskEditableViewModel = editTaskViewModel,
+                        mainViewModel = null,
                     ) {
-                        // When taskNode is Selected
-                        if (mainViewModel.selectedNode != null) {
-                            MindMapOptionsTabRow(selectedMode = selectedMode) {
-                                sheetViewModel.setMode(it)
-                            }
-
-                            // Edit Task Option
-                            // set Editing Task
-                            mainViewModel.selectedNode?.let { editTaskViewModel.setEditingTask(it) }
-                            AnimatedVisibility(
-                                visible = selectedMode == MindMapOptionsMode.EDIT_TASK,
-                                enter = slideInHorizontally(
-                                    initialOffsetX = { -width }, // small slide 300px
-                                    animationSpec = tween(
-                                        durationMillis = 200,
-                                        easing = LinearEasing // interpolator
-                                    )
-                                ),
-                                exit = ExitTransition.None
-                            ) {
-                                TaskEditContent(
-                                    taskEditableViewModel = editTaskViewModel,
-                                    mainViewModel = null,
-                                ) { dismiss() }
-                            }
-                        } else { // When mindMap Node is selected
-                            TabRow(
-                                modifier = Modifier.height(50.dp),
-                                selectedTabIndex = selectedMode.ordinal,
-                                backgroundColor = colorResource(id = R.color.transparent),
-                                contentColor = Color.LightGray,
-                            ) {
-                                Tab(
-                                    selected = true,
-                                    onClick = {},
-                                ) {
-                                    Row(verticalAlignment = Alignment.CenterVertically) {
-                                        Icon(
-                                            painter = painterResource(id = R.drawable.ic_mind_map),
-                                            contentDescription = "Edit"
-                                        )
-                                        Spacer(modifier = Modifier.width(10.dp))
-                                        Text(
-                                            "Add Child",
-                                            style = MaterialTheme.typography.subtitle1
-                                        )
-                                    }
-                                }
-                            }
-                            TaskEditContent(
-                                taskEditableViewModel = addChildViewModel,
-                                mainViewModel = null,
-                            ) { dismiss() }
-                        }
-
-                        // Add Child Option
-                        AnimatedVisibility(
-                            visible = selectedMode == MindMapOptionsMode.ADD_CHILD,
-                            enter = slideInHorizontally(
-                                initialOffsetX = { width }, // small slide 300px
-                                animationSpec = tween(
-                                    durationMillis = 200,
-                                    easing = LinearEasing // interpolator
-                                )
-                            ),
-                            exit = ExitTransition.None
-                        ) {
-                            TaskEditContent(
-                                taskEditableViewModel = addChildViewModel,
-                                mainViewModel = null,
-                            ) { dismiss() }
+                        mindMapCreteViewModel.refreshView()
+                        coroutineScope.launch { bottomSheetState.collapse() }
+                    }
+                }
+            } else { // When mindMap Node is selected
+                TabRow(
+                    modifier = Modifier.height(50.dp),
+                    selectedTabIndex = selectedMode.ordinal,
+                    backgroundColor = colorResource(id = R.color.transparent),
+                    contentColor = Color.LightGray,
+                ) {
+                    Tab(
+                        selected = true,
+                        onClick = {},
+                    ) {
+                        Row(verticalAlignment = Alignment.CenterVertically) {
+                            Icon(
+                                painter = painterResource(id = R.drawable.ic_mind_map),
+                                contentDescription = "Add child"
+                            )
+                            Spacer(modifier = Modifier.width(10.dp))
+                            Text(
+                                "Add Child",
+                                style = MaterialTheme.typography.subtitle1
+                            )
                         }
                     }
+                }
+                TaskEditContent(
+                    taskEditableViewModel = addChildViewModel,
+                    mainViewModel = null,
+                ) {
+                    mindMapCreteViewModel.refreshView()
+                    coroutineScope.launch { bottomSheetState.collapse() }
+                }
+            }
+
+            // Add Child Option
+            AnimatedVisibility(
+                visible = selectedMode == MindMapOptionsMode.ADD_CHILD,
+                enter = slideInHorizontally(
+                    initialOffsetX = { 300 }, // small slide 300px
+                    animationSpec = tween(
+                        durationMillis = 200,
+                        easing = LinearEasing // interpolator
+                    )
+                ),
+                exit = ExitTransition.None
+            ) {
+                TaskEditContent(
+                    taskEditableViewModel = addChildViewModel,
+                    mainViewModel = null,
+                ) {
+                    mindMapCreteViewModel.refreshView()
+                    coroutineScope.launch { bottomSheetState.collapse() }
                 }
             }
         }
     }
+}
 
-    /** Set position for adding child node */
-    private fun setUpAddingChildNode() {
-        mainViewModel.selectedNode?.let { selectedTask ->
-            addChildViewModel.setX(
-                (selectedTask.x ?: 0f) + 300
-            ) // set child position to right side of parent
-            addChildViewModel.setY(selectedTask.y ?: 0f)
-        } ?: run {
-            addChildViewModel.setX((mainViewModel.editingMindMap?.x ?: 0f) + 300f)
-            addChildViewModel.setY(mainViewModel.editingMindMap?.y ?: 0f)
-        }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        // update mapView with updated data
-        mindMapCreateViewModel.refreshView()
+@ExperimentalPagerApi
+@ExperimentalMaterialApi
+@ExperimentalAnimationApi
+fun setUpAddingChildNode(
+    selectedNode: Task?,
+    mainViewModel: MainViewModel,
+    addChildViewModel: AddChildViewModel,
+) {
+    selectedNode?.let { selectedTask ->
+        addChildViewModel.setX((selectedTask.x ?: 0f) + 300)
+        addChildViewModel.setY(selectedTask.y ?: 0f)
+    } ?: run {
+        addChildViewModel.setX((mainViewModel.editingMindMap?.x ?: 0f) + 300f)
+        addChildViewModel.setY(mainViewModel.editingMindMap?.y ?: 0f)
     }
 }
