--- app/src/main/res/anim/slide_in_right.xml
@@ -1,6 +1,0 @@
-<set xmlns:android="http://schemas.android.com/apk/res/android">
-    <translate android:fromXDelta="150%p" android:toXDelta="0"
-        android:duration="@android:integer/config_mediumAnimTime"/>
-    <alpha android:fromAlpha="0.0" android:toAlpha="1.0"
-        android:duration="@android:integer/config_longAnimTime" />
-</set>
