--- app/src/main/res/anim/slide_out_right.xml
@@ -1,3 +1,0 @@
-    <translate xmlns:android="http://schemas.android.com/apk/res/android"
-        android:fromXDelta="0" android:toXDelta="100%p"
-        android:duration="@android:integer/config_longAnimTime"/>
