--- app/src/main/res/layout/record_fragment.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".view.record.RecordFragment">
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-</FrameLayout>
