--- app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-<!--    <item-->
-<!--        android:id="@+id/navigation_home"-->
-<!--        android:icon="@drawable/ic_home_black_24dp"-->
-<!--        android:title="@string/title_home" />-->
-
-<!--    <item-->
-<!--        android:id="@+id/navigation_dashboard"-->
-<!--        android:icon="@drawable/ic_dashboard_black_24dp"-->
-<!--        android:title="@string/title_dashboard" />-->
-
-<!--    <item-->
-<!--        android:id="@+id/navigation_notifications"-->
-<!--        android:icon="@drawable/ic_notifications_black_24dp"-->
-<!--        android:title="@string/title_notifications" />-->
-
-    <item
-        android:id="@+id/navigation_top"
-        android:icon="@drawable/ic_top_24dp"
-        android:title="@string/top" />
-
-    <item
-        android:id="@+id/navigation_task"
-        android:icon="@drawable/ic_todo_24dp"
-        android:title="@string/task" />
-
-    <item
-        android:id="@+id/navigation_mind_map"
-        android:icon="@drawable/ic_mind_map"
-        android:title="@string/mind_map" />
-
-    <item
-        android:id="@+id/navigation_record"
-        android:icon="@drawable/ic_record_24dp"
-        android:title="@string/record" />
-
-    <item
-        android:id="@+id/navigation_settings"
-        android:icon="@drawable/ic_settings_24dp"
-        android:title="@string/settings" />
-
-</menu>
