--- app/src/main/res/navigation/main_graph.xml
@@ -1,137 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main_graph"
-    app:startDestination="@+id/navigation_top">
-
-    <fragment
-        android:id="@+id/navigation_top"
-        android:name="com.jp_funda.todomind.view.top.TopFragment"
-        android:label="@string/app_name">
-        <action
-            android:id="@+id/action_navigation_top_to_navigation_mind_map_detail"
-            app:destination="@id/navigation_mind_map_detail"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@android:anim/fade_out"
-            app:popEnterAnim="@android:anim/fade_in"
-            app:popExitAnim="@anim/slide_out_right" />
-        <action
-            android:id="@+id/action_navigation_top_to_navigation_task_detail"
-            app:destination="@id/navigation_task_detail"
-            app:enterAnim="@android:anim/fade_in"
-            app:exitAnim="@android:anim/fade_out"
-            app:popEnterAnim="@android:anim/fade_in"
-            app:popExitAnim="@android:anim/fade_out" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_task"
-        android:name="com.jp_funda.todomind.view.task.TaskFragment"
-        android:label="@string/task">
-        <action
-            android:id="@+id/action_navigation_task_to_navigation_task_detail"
-            app:destination="@id/navigation_task_detail"
-            app:enterAnim="@android:anim/fade_in"
-            app:exitAnim="@android:anim/fade_out"
-            app:popEnterAnim="@android:anim/fade_in"
-            app:popExitAnim="@android:anim/fade_out" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_task_detail"
-        android:name="com.jp_funda.todomind.view.task_detail.TaskDetailFragment">
-        <action
-            android:id="@+id/action_navigation_task_detail_to_navigation_mind_map_create"
-            app:destination="@id/navigation_mind_map_create" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_mind_map"
-        android:name="com.jp_funda.todomind.view.mind_map.MindMapFragment"
-        android:label="@string/mind_map">
-        <action
-            android:id="@+id/action_navigation_mind_map_to_navigation_mind_map_detail"
-            app:destination="@id/navigation_mind_map_detail"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@android:anim/fade_out"
-            app:popEnterAnim="@android:anim/fade_in"
-            app:popExitAnim="@anim/slide_out_right" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_mind_map_detail"
-        android:name="com.jp_funda.todomind.view.mind_map_detail.MindMapDetailFragment">
-        <action
-            android:id="@+id/action_navigation_mind_map_detail_to_navigation_task_detail"
-            app:destination="@id/navigation_task_detail" />
-        <action
-            android:id="@+id/action_navigation_mind_map_detail_to_navigation_confirm_mind_map_delete"
-            app:destination="@id/navigation_confirm_mind_map_delete" />
-        <action
-            android:id="@+id/action_navigation_mind_map_detail_to_navigation_mind_map_create"
-            app:destination="@id/navigation_mind_map_create"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@android:anim/fade_out"
-            app:popEnterAnim="@android:anim/fade_in"
-            app:popExitAnim="@anim/slide_out_right" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_mind_map_create"
-        android:name="com.jp_funda.todomind.view.mind_map_create.MindMapCreateFragment">
-        <argument
-            android:name="initialLocation"
-            android:defaultValue="@null"
-            app:argType=".view.mind_map_create.Location"
-            app:nullable="true" />
-        <action
-            android:id="@+id/action_navigation_mind_map_create_to_mindMapCreateTutorialDialog"
-            app:destination="@id/navigation_mind_map_create_tutorial" />
-    </fragment>
-
-    <dialog
-        android:id="@+id/navigation_mind_map_options_dialog"
-        android:name="com.jp_funda.todomind.view.mind_map_create.options_sheet.MindMapOptionsSheet" />
-
-    <dialog
-        android:id="@+id/navigation_confirm_mind_map_delete"
-        android:name="com.jp_funda.todomind.view.mind_map_detail.ConfirmDeleteMindMapFragment" />
-
-    <fragment
-        android:id="@+id/navigation_record"
-        android:name="com.jp_funda.todomind.view.record.RecordFragment"
-        android:label="@string/record"
-        tools:layout="@layout/record_fragment" />
-
-    <fragment
-        android:id="@+id/navigation_settings"
-        android:name="com.jp_funda.todomind.view.settings.SettingsFragment"
-        android:label="@string/settings">
-        <action
-            android:id="@+id/action_navigation_settings_to_navigation_oss_licenses"
-            app:destination="@id/navigation_oss_licenses"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@android:anim/fade_out"
-            app:popEnterAnim="@android:anim/fade_in"
-            app:popExitAnim="@anim/slide_out_right" />
-        <action
-            android:id="@+id/action_navigation_settings_to_navigation_mind_map_scale"
-            app:destination="@id/navigation_mind_map_scale" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_oss_licenses"
-        android:name="com.jp_funda.todomind.view.settings.oss_licenses.OssLicensesFragment"
-        android:label="@string/oss_licenses" />
-
-    <fragment
-        android:id="@+id/navigation_mind_map_scale"
-        android:name="com.jp_funda.todomind.view.settings.mind_map_scale.MindMapScaleFragment" />
-
-    <dialog
-        android:id="@+id/navigation_mind_map_create_tutorial"
-        android:name="com.jp_funda.todomind.view.mind_map_create.tutorial.MindMapCreateTutorialDialog"
-        android:label="MindMapCreateTutorialDialog" />
-
-</navigation>
